Update aws-vpc-cni to 1.6.0
Details
-------
AWS announced that the aws-vpc-cni plugin has been updated
to 1.6.0 and will be provided by default in EKS hosts. We
should update this chart so that it runs 1.6.0 by default.
I know there is already an open PR for this, but the author
has not addressed any of the input for weeks now.