aws-for-fluent-bit: Fix ServiceMonitor ns selector
This commit fixes a namespace selector in the Prometheus Operator's
ServiceMonitor definition. If you deploy the Helm release into, say, the
default namespace, but use a namespaceOverride to put Fluent Bit into
the "fluent-bit" or "cloudwatch" namespace, then this will look in the
wrong Kubernetes namespace for your service and you won't get metrics.